fixed attention and improved feature loading
