+++ drevalpy/models/DIPK/gene_expression_encoder.py
@@ -11,9 +11,8 @@ import torch.optim as optim
 from torch.nn import functional
 from torch.utils.data import DataLoader, Dataset
 
-ldim = 512  # 512
-hdim = [8, 4]  # [2048, 1024]
-print("Warning - using default latent dimension and hidden dimensions for gene expression autoencoder, increase dims")
+ldim = 512
+hdim = [2048, 1024]
 
 
 class GeneExpressionEncoder(nn.Module):
@@ -152,8 +151,7 @@ def train_gene_expession_autoencoder(gene_expression_input: np.ndarray) -> GeneE
     """
     lr = 1e-4
     batch_size = 1024
-    epochs = 1
-    print("Warning - training gene expression autoencoder for 1 epoch only")
+    epochs = 100
     noising = True
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
